OCSADV-402 Move BandsList to spModel.core #time 15m
