+++ bundle/edu.gemini.ags/src/main/scala/edu/gemini/ags/api/AgsStrategy.scala
@@ -2,10 +2,10 @@ package edu.gemini.ags.api
 
 import edu.gemini.ags.api.AgsAnalysis.NotReachable
 import edu.gemini.ags.api.AgsMagnitude.{MagnitudeCalc, MagnitudeTable}
-import edu.gemini.catalog.api.{BandsList, CatalogQuery}
+import edu.gemini.catalog.api.CatalogQuery
 import edu.gemini.pot.ModelConverters._
 import edu.gemini.spModel.ags.AgsStrategyKey
-import edu.gemini.spModel.core.Angle
+import edu.gemini.spModel.core.{BandsList, Angle}
 import edu.gemini.spModel.core.Target.SiderealTarget
 import edu.gemini.spModel.guide.{ValidatableGuideProbe, GuideProbe}
 import edu.gemini.spModel.obs.context.ObsContext
@@ -46,7 +46,9 @@ trait AgsStrategy {
 
   def guideProbes: List[GuideProbe]
 
-  // Return the bands searched by the strategy
+  /**
+   * Indicates the bands that will be used for a given probe
+   */
   def probeBands: BandsList
 }
