+++ bundle/edu.gemini.ags/src/main/scala/edu/gemini/ags/impl/ScienceTargetStrategy.scala
@@ -1,10 +1,10 @@
 package edu.gemini.ags.impl
 
 import edu.gemini.ags.api.AgsMagnitude._
-import edu.gemini.catalog.api.BandsList
 import edu.gemini.pot.ModelConverters._
 import edu.gemini.ags.api.{AgsAnalysis, AgsStrategy}
 import edu.gemini.spModel.ags.AgsStrategyKey
+import edu.gemini.spModel.core.BandsList
 import edu.gemini.spModel.core.Target.SiderealTarget
 import edu.gemini.spModel.guide.{ValidatableGuideProbe, GuideProbe}
 import edu.gemini.spModel.obs.context.ObsContext
