+++ bundle/edu.gemini.ags/src/main/scala/edu/gemini/ags/impl/Strategy.scala
@@ -1,12 +1,11 @@
 package edu.gemini.ags.impl
 
 import edu.gemini.ags.api.AgsStrategy
-import edu.gemini.catalog.api.NiciBandsList
 import edu.gemini.catalog.votable.RemoteBackend
 import edu.gemini.pot.sp.SPComponentType
 import edu.gemini.spModel.ags.AgsStrategyKey
 import edu.gemini.spModel.ags.AgsStrategyKey._
-import edu.gemini.spModel.core.Site
+import edu.gemini.spModel.core.{NiciBandsList, Site}
 import edu.gemini.spModel.gemini.altair.{AltairParams, InstAltair}
 import edu.gemini.spModel.gemini.gems.Canopus
 import edu.gemini.spModel.gemini.nici.NiciOiwfsGuideProbe
