+++ bundle/edu.gemini.catalog/src/test/scala/edu/gemini/catalog/api/MagnitudeConstraintsSpec.scala
@@ -2,7 +2,7 @@ package edu.gemini.catalog.api
 
 import edu.gemini.catalog.api.MagnitudeLimits.{SaturationLimit, FaintnessLimit}
 import edu.gemini.spModel.core.Target.SiderealTarget
-import edu.gemini.spModel.core.{ Coordinates, Magnitude, MagnitudeBand}
+import edu.gemini.spModel.core._
 import org.specs2.mutable.SpecificationWithJUnit
 
 class MagnitudeConstraintsSpec extends SpecificationWithJUnit {
